+++ Changelog.rst
@@ -1,3 +1,13 @@
+Version NEXTVERSION
+----------------
+
+**2026-03-??**
+
+* Fix `cfdm.write` for the writing of coordinates with formula_terms
+  (https://github.com/NCAS-CMS/cfdm/issues/???)
+
+----
+
 Version 1.13.0.0
 ----------------